# openapi.yaml
openapi: 3.1.0
info:
  title: ECHO Platform API
  version: 1.0.0
  description: |
    ECHO Platform REST API and WebSocket specification.
    
    **Architecture**: Client-server relay with E2E encryption. The API server is a stateless
    relay and operational cache. It cannot read message content (E2E encrypted blobs only),
    does not own user identities (holders use did:key), and does not control token balances
    (Constellation metagraph).
    
    **Authentication**: All endpoints except /v1/auth/* require a Bearer token in the
    Authorization header. Tokens are issued after passkey/biometric verification.
    
    **Encryption**: Request/response bodies for sensitive endpoints are encrypted with
    a session key (AES-256-GCM) negotiated during auth. Message content is always E2E
    encrypted on the client (X25519 + ChaCha20-Poly1305) before transmission.
    
    **On-chain references**: Endpoints that interact with blockchain state (tokens, identity,
    trust) return cached values with a `snapshotHash` field for client-side verification
    against on-chain state (Phase 3+).
  contact:
    name: ECHO Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.echo.network/v1
    description: Production
  - url: https://api.staging.echo.network/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /identity/did:
    post:
      tags: [Identity]
      summary: (Legacy Phase-0) Create a chain-anchored DID
      description: |
        Deprecated. Phase 1 (ADR-0001) replaced chain-anchored DIDs with
        self-certifying did:key. New clients should use POST /identity/register.
        This endpoint is retained for backward compatibility with pre-Phase-1
        iOS builds and will be removed once WO-280 ships.
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [publicKey]
              properties:
                publicKey:
                  type: string
                  format: base64url
                  description: P-256 public key from Secure Enclave
      responses:
        "202":
          description: DID creation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DIDDocument"
  /identity/did/{did}:
    get:
      tags: [Identity]
      summary: (Legacy Phase-0) Resolve a chain-anchored DID document
      description: Deprecated. Use GET /identity/{did} for did:key bindings.
      deprecated: true
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DID document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DIDDocument"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from /auth/verify
...
  schemas:
...
    DIDDocument:
      type: object
      properties:
        did:
          type: string
          description: "W3C did:key identifier. Phase 1 — see ADR-0001."
          example: "did:key:z2dmzD81cgPx8Vki7JbuuMmFYrWPgYoytykUZ3eyqht1j9KbrL2g"
        publicKeys:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
                enum: [EcdsaSecp256r1VerificationKey2019, X25519KeyAgreementKey2020]
              publicKeyBase64:
                type: string
                format: base64url
        serviceEndpoints:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              endpoint:
                type: string
                format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        l1TxHash:
          type: string
          description: Optional Identity Metagraph L1 anchor / transaction reference for verification
...